+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Features:
 
-  - Added CHANGELOG.md generation at the code_generator::cookbook
+  - Added CHANGELOG.md, Vagrantfile generation at the code_generator::cookbook
   - Fixed the supporting platform to `centos-7.0`
 
 ## 0.1.0